requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bnbphoneticparser==0.1.5
boto3==1.14.58
botocore==1.17.58
bs4==0.0.1
cachetools==4.2.0
cattrs==1.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
contextvars==2.4
crayons==0.3.1
cssselect==1.1.0
dataclasses==0.7
dataclasses-json==0.5.2
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
et-xmlfile==1.0.1
facebook-scraper==0.2.17
facebook-sdk==3.1.0
fake-useragent==0.1.11
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-trans-new==1.1.9
googletrans==3.0.0
grpcio==1.32.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.9.2
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
immutables==0.14
importlib-metadata==3.3.0
instagram-scraper==1.9.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-enum==1.5.1
mistune==0.8.4
MouseInfo==0.1.3
moviepy==1.0.3
munkres==1.1.2
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.1
pandocfilters==1.4.3
panphon==0.17
parse==1.18.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.7
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycld3==0.20
pycparser==2.20
pyee==7.0.4
PyGetWindow==0.0.9
Pygments==2.6.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pyppeteer==0.2.2
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.14.0
python-facebook-api==0.6.1
python-youtube==0.6.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
responses==0.12.0
rfc3986==1.4.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.1
sqlparse==0.4.1
stringcase==1.2.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
textblob==0.15.3
tokenizers==0.9.4
tornado==6.0.4
tqdm==4.58.0
traitlets==4.3.3
transformers==4.0.0
typing==3.7.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.10
vaderSentiment==3.3.2
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
